% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{luxembourg}
\alias{luxembourg}
\title{Numbers from Luxembourg's official statistics on governmental deficit, surplus and debt from 1999 to 2019}
\format{
A numeric vector of length 496.
}
\source{
\url{https://ec.europa.eu/eurostat/data/database}
}
\usage{
luxembourg
}
\description{
Compilation of all the numbers from Luxembourg's official statistics on governmental deficit, surplus, debt and associated accounts from 1999 to 2019.
}
\details{
Data was extracted from \href{https://ec.europa.eu/eurostat/data/database}{Eurostat's database} in August 2020 through the directory: Database by themes -> Economy and finance -> Government statistics -> Government finance statistics -> Government deficit and debt -> Government deficit/surplus, debt and associated data. After sub-setting by country and selecting only data from 1999 to 2019, all the numbers from all the tables in the aforementioned category were pooled together.
}
\keyword{datasets}
