% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{luxembourg_bl2}
\alias{luxembourg_bl2}
\title{Second digits from Luxembourg's official statistics on government deficit, surplus and debt from 1999 to 2019.}
\format{
A numeric vector of length 464.
}
\source{
\url{https://ec.europa.eu/eurostat/data/database}
}
\usage{
luxembourg_bl2
}
\description{
Second digits from Luxembourg's official statistics on government deficit, surplus, debt and associated data from 1999 to 2019.
}
\details{
Data was extracted from \href{https://ec.europa.eu/eurostat/data/database}{Eurostat's database} in August 2020 through the directory: Database by themes -> Economy and finance -> Government statistics -> Government finance statistics -> Government deficit and debt -> Government deficit/surplus, debt and associated data. After sub-setting by country and selecting only data from 1999 to 2019, all the numbers from all the tables in this category were aggregated and then \code{\link[daubl]{smsdigit}} was used to obtain the second digits.
}
\keyword{datasets}
