% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{smsdigit}
\alias{smsdigit}
\title{The second most significant digits of numbers}
\usage{
smsdigit(x)
}
\arguments{
\item{x}{A numeric vector or matrix.}
}
\value{
\code{smsdigit} returns a numeric vector with {\link[base]{length}} equal to the number of elements of \code{x} that are greater than or equal to 10 plus the number of \code{NA} values in \code{x}.
}
\description{
Obtain the second most significant digits of numbers in numeric vectors or matrices.
}
\details{
For each non-zero, non-missing element of \code{x} that is greater than or equal to 10, \code{smsdigit} returns the second digit of that number's floating point representation of base 10. \code{NA} values in \code{x} generate \code{NA} values in the output of \code{smsdigit}.
}
\examples{
smsdigit(451)
smsdigit(0.00451)
smsdigit(-0.00451)
smsdigit(c(-31.12, -0.23, 0.1234, 5.32))

x <- c(11, 22, 33, NA)
length(x)
length(smsdigit(x)) # equal to length(x)

x <- c(0, 1, 2,  10, 20)
length(x)
length(smsdigit(x)) # not equal to length(x)

}
\seealso{
\itemize{
\item \code{\link[daubl]{msdigit}} to obtain the most significant digits of numbers.
}
}
