% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{benford}
\alias{benford}
\title{Benford's law for first and second digits}
\usage{
benford(d)
}
\arguments{
\item{d}{Numeric vector of {\link[base]{length}} 1. Admissible values: 1 or 2.}
}
\value{
\itemize{
\item \code{benford(1)} returns a numeric vector of {\link[base]{length}} 9.
\item \code{benford(2)} returns a numeric vector {\link[base]{length}} 10.
}
}
\description{
First and second digit probabilities, as postulated by Benford's law.
}
\details{
Benford’s law is an empirical regularity regarding the occurrence of digits in numbers that was first introduced by \insertCite{newcomb1881note;textual}{daubl} and later popularized and supported with empirical evidence by \insertCite{benford1938;textual}{daubl}. The formulas that give Benford's law first and second digit probabilities can be consulted in \insertCite{hill1995derivation;textual}{daubl}.
}
\examples{
benford(1)
benford(2)

}
\references{
\insertAllCited{}
}
\seealso{
\itemize{
\item \code{\link[daubl]{msdigit}} to obtain the most significant digits of numbers.
\item \code{\link[daubl]{smsdigit}} to obtain the second most significant digits of numbers.
}
}
